requirement.txt
django==2.1.9
Pillow==6.0.0
django-tinymce4-lite==1.7.5
django-filebrowser-no-grappelli==3.7.8
sorl-thumbnail==12.5.0
selenium==3.141.0


